+++ croaster/src/CroasterWiFiManager.cpp
@@ -1,0 +1,49 @@
+#include "CroasterWiFiManager.h"
+#include "CroasterConstants.h"
+#if defined(ESP32)
+#include <WiFi.h>
+#elif defined(ESP8266)
+#include <ESP8266WiFi.h>
+#endif
+
+static WiFiManager wifiManager;
+
+void CroasterWiFiManager::configModeCallback(WiFiManager *myWiFiManager)
+{
+    debugln("# Config mode: " + WiFi.softAPIP().toString());
+}
+
+void CroasterWiFiManager::setup(const String &apName)
+{
+    debugln("# Setting up WiFi Manager");
+
+    WiFi.mode(WIFI_STA);
+
+    wifiManager.setDebugOutput(true);
+    wifiManager.setConfigPortalBlocking(false);
+    wifiManager.setAPCallback(configModeCallback);
+    wifiManager.setAPStaticIPConfig(IPAddress(10, 0, 1, 1), IPAddress(10, 0, 1, 1), IPAddress(255, 255, 255, 0));
+    wifiManager.setClass("invert");
+    wifiManager.setConnectTimeout(10);
+
+    if (wifiManager.autoConnect(apName.c_str()))
+    {
+        debugln("# WiFi Connected");
+    }
+}
+
+void CroasterWiFiManager::process()
+{
+    wifiManager.process();
+}
+
+void CroasterWiFiManager::restart()
+{
+    wifiManager.reboot();
+}
+
+void CroasterWiFiManager::erase()
+{
+    wifiManager.erase();
+    restart();
+}